Parser optimization
Optimized the parser by removing the need to iterate through the delinators. I swapped the vector of delinators for an unordered_map to allow for quick access. I also used the secondary value of the unordered_map to store a delinator class. This delinator class allows for me to specify that anything after a hashtag(the comment delinator) should be ignored without the need to iterate through the output strings later to remove the comments.